+++ src/actions/wallet/signAuthorization.test.ts
@@ -380,7 +380,7 @@ test('error: unsupported account type', async () => {
 
     The \`signAuthorization\` Action does not support JSON-RPC Accounts.
 
-    Docs: https://viem.sh/experimental/eip7702/signAuthorization
+    Docs: https://viem.sh/docs/eip7702/signAuthorization
     Version: viem@x.y.z]
   `)
 })
@@ -397,7 +397,7 @@ test('error: no account', async () => {
     [AccountNotFoundError: Could not find an Account to execute with this Action.
     Please provide an Account with the \`account\` argument on the Action, or by supplying an \`account\` to the Client.
 
-    Docs: https://viem.sh/experimental/eip7702/signAuthorization
+    Docs: https://viem.sh/docs/eip7702/signAuthorization
     Version: viem@x.y.z]
   `)
 })